updated for version 7.2b-000

+++ src/message.c
@@ -1390,7 +1390,7 @@ msg_outtrans_len_attr(msgstr, len, attr)
 									attr);
 		plain_start = str + 1;
 		msg_puts_attr(s, attr == 0 ? hl_attr(HLF_8) : attr);
-		retval += STRLEN(s);
+		retval += (int)STRLEN(s);
 	    }
 	    else
 		++retval;
